requirements.txt
Flask==3.0.3
google-generativeai==0.8.3
google-auth==2.29.0
python-dotenv==1.0.1
gunicorn==22.0.0
